# Conflicts:
#	STM/MainController/Inc/control/lst_distance.h
#	STM/MainController/Inc/lst_obstacle.h
#	STM/MainController/Src/control/lst_distance.c
#	STM/MainController/Src/control/lst_movement.c
#	STM/MainController/Src/lst_bt.c
#	STM/MainController/Src/lst_control.c
#	STM/MainController/Src/lst_obstacle.c

+++ STM/MainController/Src/lst_utils.c
@@ -20,8 +20,8 @@ float interpolate_return[2];
 /**
  * @brief Copies len number of bytes from pSrc to pDest
  */
-void LST_Utils_Memory_Copy(uint8_t *pDest, uint8_t *pSrc, uint8_t len) {
-  uint8_t i;
+void LST_Utils_Memory_Copy(uint8_t *pDest, uint8_t *pSrc, uint16_t len) {
+  uint16_t i;
   // Manually copy the data
   for (i = 0; i < len; i++) {
     // Copy data from source to destination
